requirements.txt
Django>=4.1.7
django-crispy-forms>=2.0
Pillow>=9.4.0
psycopg2>=2.9.5